Safari-611.1.10/LayoutTests/http/tests/security/xss-DENIED-xsl-document-securityOrigin-expected.txt

CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
This test passes if it doesn't alert the contents of innocent-victim.html.
